feat: swap node-fetch for undici
node-fetch does not use browser streams which makes using it in typed
environments very hard since the methods on node streams are different
from those on browser streams.

undici is a node implementation of fetch that supports browser streams
and is probably the closest thing we'll ever see to fetch in node core,
so swap node-fetch for undici.

BREAKING CHANGE: only browser streams are returned, requires node 16+